+++ src/config.h
@@ -1,6 +1,8 @@
 #ifndef CONFIG_H
 #define CONFIG_H
 
+#include <stddef.h>
+
 typedef struct {
     char *name;
     char *model;
@@ -38,12 +40,12 @@ typedef struct {
 
 /* Load config from ~/.artifice/config.yaml and ./.artifice/config.yaml.
  * Returns 0 on success. errbuf receives error message on failure. */
-int config_load(config_t *cfg, char *errbuf, int errlen);
+int config_load(config_t *cfg, char *errbuf, size_t errlen);
 void config_free(config_t *cfg);
 
 /* Resolve agent by name (NULL = default). Returns 0 on success. */
 int resolve_agent(const config_t *cfg, const char *name,
-                  resolved_agent_t *out, char *errbuf, int errlen);
+                  resolved_agent_t *out, char *errbuf, size_t errlen);
 void resolved_agent_free(resolved_agent_t *ra);
 
 /* Install default config to ~/.artifice/ */
